+++ components/detail/page.module.scss
@@ -220,32 +220,6 @@ $backColor: rgba(215, 196, 255, 0.179);
   flex-direction: column;
 }
 
-//글작성 페이지 헤더
-.writeHeader {
-  padding: 1rem;
-  text-align: center;
-  position: relative;
-  h1 {
-    text-align: center;
-    font-size: var(--font-header-l);
-  }
-  button {
-    padding: 0.5rem;
-    border: none;
-    border-radius: var(--border-radius-s);
-    background-color: rgb(243, 237, 255);
-    cursor: pointer;
-    &:hover {
-      background-color: rgb(163, 122, 253);
-      color: #fff;
-    }
-  }
-  .backBtn {
-    position: absolute;
-    left: 0;
-  }
-}
-
 /* 데스크탑 화면 스타일 */
 @media (min-width: 768px) {
   .container {
